Roll the default context builder into contextbuilder, removing an unutilized level of indirection
