+++ pi4j-core/src/main/java/com/pi4j/context/ContextBuilder.java
@@ -1,8 +1,14 @@
 package com.pi4j.context;
 
+import com.pi4j.boardinfo.util.BoardInfoHelper;
 import com.pi4j.config.Builder;
-import com.pi4j.context.impl.DefaultContextBuilder;
+import com.pi4j.context.impl.DefaultContext;
+import com.pi4j.exception.Pi4JException;
 import com.pi4j.provider.Provider;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import java.util.*;
 
 /**
  * Fluent builder used to configure and create a Pi4J {@link Context}. It accumulates settings such as
@@ -14,15 +20,36 @@ import com.pi4j.provider.Provider;
  * @see Context
  * @see ContextConfig
  */
-public interface ContextBuilder extends Builder<Context> {
+public class ContextBuilder implements Builder<Context> {
+    protected Logger logger = LoggerFactory.getLogger(ContextBuilder.class);
+
+    // auto detection flags
+    protected boolean autoDetectMockPlugins = !BoardInfoHelper.runningOnRaspberryPi();
+    protected boolean autoDetectPlatforms = false;
+    protected boolean autoDetectProviders = false;
+    protected boolean enableShutdownHook = false;
+
+    // default platform identifier
+    protected String defaultPlatformId = null;
+
+    // extensibility modules
+    protected Collection<Provider> providers = Collections.synchronizedList(new ArrayList<>());
+
+    // properties
+    protected Map<String,String> properties = Collections.synchronizedMap(new HashMap<>());
+
+    // Add a field to store the GPIO chip name
+    protected String gpioChipName;
 
     /**
-     * Creates a new, empty context builder.
-     *
-     * @return a fresh {@link ContextBuilder} instance
+     * Private Constructor
      */
-    static ContextBuilder newInstance(){
-        return DefaultContextBuilder.newInstance();
+    private ContextBuilder() {
+        // forbid object construction
+    }
+
+    public static ContextBuilder newInstance(){
+        return new ContextBuilder();
     }
 
     /**
@@ -31,7 +58,11 @@ public interface ContextBuilder extends Builder<Context> {
      * @param provider the providers to add
      * @return this builder instance for method chaining
      */
-    ContextBuilder add(Provider ... provider);
+    public ContextBuilder add(Provider... provider) {
+        if(provider != null && provider.length > 0)
+            this.providers.addAll(List.of(provider));
+        return this;
+    }
 
     /**
      * Enables auto-detection of mock plugins on the classpath, primarily useful for testing without real
@@ -39,35 +70,50 @@ public interface ContextBuilder extends Builder<Context> {
      *
      * @return this builder instance for method chaining
      */
-    ContextBuilder autoDetectMockPlugins();
+    public ContextBuilder autoDetectMockPlugins() {
+        this.autoDetectMockPlugins = true;
+        return this;
+    }
 
     /**
      * Enables auto-detection of platform implementations available on the classpath.
      *
      * @return this builder instance for method chaining
      */
-    ContextBuilder autoDetectPlatforms();
+    public ContextBuilder autoDetectPlatforms() {
+        this.autoDetectPlatforms = true;
+        return this;
+    }
 
     /**
      * Disables auto-detection of platform implementations on the classpath.
      *
      * @return this builder instance for method chaining
      */
-    ContextBuilder noAutoDetectPlatforms();
+    public ContextBuilder noAutoDetectPlatforms() {
+        this.autoDetectPlatforms = false;
+        return this;
+    }
 
     /**
      * Enables auto-detection of provider implementations available on the classpath.
      *
      * @return this builder instance for method chaining
      */
-    ContextBuilder autoDetectProviders();
+    public ContextBuilder autoDetectProviders() {
+        this.autoDetectProviders = true;
+        return this;
+    }
 
     /**
      * Disables auto-detection of provider implementations on the classpath.
      *
      * @return this builder instance for method chaining
      */
-    ContextBuilder noAutoDetectProviders();
+    public ContextBuilder noAutoDetectProviders() {
+        this.autoDetectProviders = false;
+        return this;
+    }
 
     /**
      * Enables registration of a JVM shutdown hook that automatically shuts the context down when the JVM
@@ -75,35 +121,21 @@ public interface ContextBuilder extends Builder<Context> {
      *
      * @return this builder instance for method chaining
      */
-    ContextBuilder enableShutdownHook();
-
-    /**
-     * Disables registration of the JVM shutdown hook, leaving the caller responsible for shutting the
-     * context down explicitly.
-     *
-     * @return this builder instance for method chaining
-     */
-    ContextBuilder disableShutdownHook();
-
-    /**
-     * Enables or disables the JVM shutdown hook depending on the given flag.
-     *
-     * @param enableShutdownHook {@code true} to register the shutdown hook, {@code false} to skip it
-     * @return this builder instance for method chaining
-     */
-    default ContextBuilder setShutdownHook(boolean enableShutdownHook) {
-        if (enableShutdownHook)
-            return enableShutdownHook();
-        else
-            return disableShutdownHook();
+    public ContextBuilder enableShutdownHook() {
+        this.enableShutdownHook = true;
+        return this;
     }
 
-    /**
-     * Builds an immutable configuration snapshot from the current builder state, without creating a context.
-     *
-     * @return a {@link ContextConfig} reflecting this builder's settings
-     */
-    ContextConfig toConfig();
+    public ContextBuilder disableShutdownHook() {
+        this.enableShutdownHook = false;
+        return this;
+    }
+
+    public ContextBuilder setGpioChipName(String chipName) {
+        this.gpioChipName = chipName;
+        return this;
+    }
+
 
     /**
      * Enables or disables auto-detection of both platforms and providers depending on the given flag.
@@ -111,7 +143,7 @@ public interface ContextBuilder extends Builder<Context> {
      * @param autoDetect {@code true} to enable auto-detection, {@code false} to disable it
      * @return this builder instance for method chaining
      */
-    default ContextBuilder setAutoDetect(boolean autoDetect){
+    public ContextBuilder setAutoDetect(boolean autoDetect){
         if(autoDetect)
             return autoDetect();
         else
@@ -123,10 +155,10 @@ public interface ContextBuilder extends Builder<Context> {
      *
      * @return this builder instance for method chaining
      */
-    default ContextBuilder autoDetect(){
+    public ContextBuilder autoDetect(){
         // auto detect all extensibility modules in the classpath
         return  autoDetectPlatforms().
-                autoDetectProviders();
+            autoDetectProviders();
     }
 
     /**
@@ -134,17 +166,77 @@ public interface ContextBuilder extends Builder<Context> {
      *
      * @return this builder instance for method chaining
      */
-    default ContextBuilder noAutoDetect(){
+    public ContextBuilder noAutoDetect(){
         // do not auto detect all extensibility modules in the classpath
         return  noAutoDetectPlatforms().
-                noAutoDetectProviders();
+            noAutoDetectProviders();
     }
 
     /**
-     * Sets the GPIO chip name to be used by the resulting context.
+     * Enables or disables the JVM shutdown hook depending on the given flag.
      *
-     * @param chipName the name of the GPIO chip (e.g., "gpiochip0")
+     * @param enableShutdownHook {@code true} to register the shutdown hook, {@code false} to skip it
      * @return this builder instance for method chaining
      */
-    ContextBuilder setGpioChipName(String chipName);
+    ContextBuilder setShutdownHook(boolean enableShutdownHook) {
+        if (enableShutdownHook)
+            return enableShutdownHook();
+        else
+            return disableShutdownHook();
+    }
+
+    /**
+     * Builds an immutable configuration snapshot from the current builder state, without creating a context.
+     *
+     * @return a {@link ContextConfig} reflecting this builder's settings
+     */
+    public ContextConfig toConfig() {
+
+        // create a new context configuration object
+        return new ContextConfig() {
+            private final Collection<Provider> providers = Collections.unmodifiableCollection(
+                new ArrayList<>(ContextBuilder.this.providers));
+            private final boolean autoDetectMockPlugins = ContextBuilder.this.autoDetectMockPlugins;
+            private final boolean autoDetectPlatforms = ContextBuilder.this.autoDetectPlatforms;
+            private final boolean enableShutdownHook = ContextBuilder.this.enableShutdownHook;
+            private final boolean autoDetectProviders = ContextBuilder.this.autoDetectProviders;
+
+            @Override
+            public Collection<Provider> providers() {
+                return providers;
+            }
+
+            @Override
+            public boolean autoDetectMockPlugins() {
+                return autoDetectMockPlugins;
+            }
+
+            @Override
+            public boolean autoDetectPlatforms() {
+                return autoDetectPlatforms;
+            }
+
+            @Override
+            public boolean enableShutdownHook() {
+                return enableShutdownHook;
+            }
+
+            @Override
+            public boolean autoDetectProviders() {
+                return autoDetectProviders;
+            }
+        };
+    }
+
+    @Override
+    public Context build() throws Pi4JException {
+        logger.trace("invoked 'build()'");
+
+        // create new context
+        Context context = DefaultContext.newInstance(this.toConfig());
+
+        // return the newly created context
+        logger.debug("Pi4J successfully created and initialized a new runtime 'Context'.'");
+        return context;
+    }
 }
